PSquare: Implement `__serialize()` and `__unserialize()`
Avoid unsafe serialization problems in the future by offering a
stable serialization format.

The extra serialize/unserialize methods exist only for PHP 7.2/7.3,
as PHP 7.4+ will use __serialize()/__unserialize() instead without
needing to manually serialize the data.
https://www.php.net/manual/en/language.oop5.magic.php#object.serialize
https://wiki.php.net/rfc/phase_out_serializable

This is ugly and repetitive code to have in the library. I considered
making our own toArray() and fromArray() instead. That would be neater
in the source here, but would require non-standard discovery and usage
in callers, and we would likely then deprecate that in a year in
favour of native __serialize, and then release a API-breaking semver
major version of the library, and update callers. We can avoid all that
by taking a bit of redundant code in the library. Consumers can then
easily support both this and the next version (after PHP 7.4+ semver
bump), e.g. MediaWiki can tolerate `running-stat ^2.0.0 || ^3.0.0`
and the same for any other dependants, thus containing the overhead
to this library's own code only.

Bug: T313341
Change-Id: Idf0072c7eb7731ace9e69f4e6e0c297f5e841b62